portfolio-optimizer
This is the main code for simulating an investment portfolio buy/sell strategy. The purpose of this software is to eventually make a framework to easily test buy/sell algorithms across many different possible holdings in a portfolio simultaneously. When the code is fast enough, Monte Carlo simulations can then be run using this framework to see how buying strategies would have worked historically, for example, using random stock subsets or time period subsets. 

I feel the main concerns at the moment should be to work on speeding up the code, and then making it easier to update the tested buy/sell strategy. It would also be good to include an automated benchmarking tool. (Perhaps one that compares the simulation's returns versus the returns of the S&P over the same time period.) Once those issues are taken care of, we should work on user interface to practice releasing a finished product.